# Optimize Curl multiplexing