# Description of change

This PR avoids conditional branching during multiplexing and de-multiplexing.

This achieves a 2x speedup for batched Curl on amd64 over current `master`:
```
name                old time/op  new time/op  delta
CurlTransaction     6.03ms ± 3%  3.49ms ± 6%  -42.17%
CurlHash             299µs ± 4%   132µs ± 6%  -56.05%
```
It also achieves a minor speedup for the regular Curl on amd64 over current `master`:
```
name                old time/op  new time/op  delta
CurlTransaction      338µs ± 4%   339µs ± 6%     ~
CurlHash            12.5µs ± 6%  10.9µs ± 4%  -12.86%
```

## Type of change

- Enhancement (a non-breaking change which adds functionality)

## Change checklist

- [x] My code follows the contribution guidelines for this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] New and existing unit tests pass locally with my changes